Decompose daemon control file
