+++ daemon/control_shutdown.go
@@ -1,0 +1,180 @@
+package daemon
+
+import (
+	"errors"
+	"fmt"
+	"io/fs"
+	"net/rpc"
+	"os"
+	"strings"
+	"syscall"
+	"time"
+)
+
+// ShutdownResult reports how RequestShutdown stopped (or failed to stop) the
+// running daemon. Used by upgrade.go and autoupdate.go to pick the right
+// user-facing message after a binary swap.
+type ShutdownResult int
+
+const (
+	// ShutdownNoDaemon means no daemon was running (no socket, ECONNREFUSED,
+	// or PID-file scan found nothing). The upgrade prints bare success.
+	ShutdownNoDaemon ShutdownResult = iota
+	// ShutdownViaRPC means the daemon acknowledged the Shutdown RPC and is
+	// exiting cleanly. The post-#501 happy path.
+	ShutdownViaRPC
+	// ShutdownViaSIGTERM means the daemon was a pre-#501 binary that did not
+	// register the Shutdown RPC, so we located its PID and signaled it
+	// directly. The upgrade prints a slightly different success message so
+	// users know we used the fallback. See #504.
+	ShutdownViaSIGTERM
+	// ShutdownFailed means a daemon was proven to be listening on the
+	// control socket (the Shutdown RPC came back as method-not-found, not
+	// ECONNREFUSED) but the SIGTERM fallback could not locate a PID to
+	// signal — e.g. no PID file AND pgrep is unavailable on this host. The
+	// daemon is still running the old binary; the caller must surface the
+	// recovery hint in the accompanying error. See #553.
+	ShutdownFailed
+	// ShutdownError means the control socket was present and a Shutdown RPC
+	// was attempted, but it failed with an error that does NOT prove the
+	// daemon absent and is NOT method-not-found: EACCES (socket exists but
+	// the caller lacks permission to connect), ECONNRESET/EPIPE (the
+	// connection was established then reset), or a dial timeout (the socket
+	// is bound but the listener is unresponsive). All of these imply a daemon
+	// WAS listening, so reporting ShutdownNoDaemon — documented as "no daemon
+	// was running" — would mislabel the outcome. The daemon's final state is
+	// unknown and it may still be running; the accompanying error carries the
+	// detail. See #978.
+	ShutdownError
+)
+
+// sigtermFallbackGrace is the max time we wait for a SIGTERM'd daemon to exit
+// before escalating to SIGKILL.
+const sigtermFallbackGrace = 5 * time.Second
+
+// sigtermFallbackPoll is how often we check whether the SIGTERM'd daemon has
+// exited.
+const sigtermFallbackPoll = 100 * time.Millisecond
+
+// RequestShutdown asks any running daemon to exit cleanly. The normal path
+// uses the Shutdown RPC (#498/#501). When the running daemon is a pre-#501
+// binary that does not register Shutdown, we fall back to locating the
+// daemon's PID and sending SIGTERM directly (#504) so an `af upgrade` does
+// not leave a stale daemon running the old binary.
+//
+// Returns (ShutdownNoDaemon, nil) when no daemon is running (no socket or
+// ECONNREFUSED), (ShutdownViaRPC, nil) when the Shutdown RPC acknowledged,
+// (ShutdownViaSIGTERM, nil) when the fallback signaled a real `af --daemon`
+// process, (ShutdownFailed, err) when the daemon is provably running but
+// the fallback could not locate or signal it (ambiguous pgrep matches, no
+// PID file with pgrep unavailable, permission denied on signal) — the
+// returned error carries the recovery hint the caller must surface — and
+// (ShutdownError, err) when the socket was present but the Shutdown RPC
+// failed with a transport error that is neither daemon-absent nor
+// method-not-found (EACCES, ECONNRESET/EPIPE, dial timeout): a daemon was
+// listening but its final state is unknown (#978).
+func RequestShutdown() (ShutdownResult, error) {
+	socketPath, err := DaemonSocketPath()
+	if err != nil {
+		return ShutdownNoDaemon, err
+	}
+	if _, statErr := os.Stat(socketPath); statErr != nil {
+		if errors.Is(statErr, fs.ErrNotExist) {
+			return ShutdownNoDaemon, nil
+		}
+		return ShutdownNoDaemon, statErr
+	}
+	var resp ShutdownResponse
+	if rpcErr := callDaemonNoEnsure("Shutdown", ShutdownRequest{}, &resp); rpcErr != nil {
+		if isDaemonAbsentErr(rpcErr) {
+			return ShutdownNoDaemon, nil
+		}
+		if isRPCMethodNotFoundErr(rpcErr) {
+			// Daemon is alive on the socket but does not speak Shutdown
+			// (pre-#501 binary). Fall through to the PID-based fallback.
+			return sigtermFallback()
+		}
+		// The socket was present (os.Stat above succeeded) and the error is
+		// neither daemon-absent (ECONNREFUSED/ENOENT) nor method-not-found:
+		// EACCES, ECONNRESET/EPIPE, or a dial timeout. Something was listening,
+		// so ShutdownNoDaemon would mislabel this — report the ambiguous
+		// contacted-but-errored outcome instead (#978).
+		return ShutdownError, rpcErr
+	}
+	if !resp.OK {
+		return ShutdownNoDaemon, fmt.Errorf("daemon Shutdown RPC returned OK=false")
+	}
+	return ShutdownViaRPC, nil
+}
+
+// shutdownCompleteGrace bounds how long WaitForShutdownCompletion polls for
+// the control socket to stop answering; shutdownCompletePoll is the cadence.
+// Package vars rather than constants so tests can shorten the timeout path,
+// mirroring stopDaemonGrace/stopDaemonPoll. The grace matches
+// sigtermFallbackGrace — the wait signalAndWait already imposes on the
+// SIGTERM path. The poll is tighter than sigtermFallbackPoll because the
+// normal RPC teardown completes just past shutdownAckGrace (50ms), so a 50ms
+// cadence usually resolves the wait on its first or second check.
+var (
+	shutdownCompleteGrace = sigtermFallbackGrace
+	shutdownCompletePoll  = shutdownAckGrace
+)
+
+// WaitForShutdownCompletion blocks until the daemon control socket stops
+// answering pings, bounded by shutdownCompleteGrace. The Shutdown RPC
+// acknowledges before the daemon tears down (shutdownAckGrace plus the
+// teardown tail), so a caller that respawns immediately after RequestShutdown
+// races the dying daemon: EnsureDaemon's liveness ping — or a unit-restarted
+// daemon's startup ping guard — can see the old socket still answering, skip
+// the spawn, and leave nothing running once the old daemon exits (#854).
+// Callers on the shutdown-then-respawn path must wait for this to return
+// before respawning. It mirrors signalAndWait's poll-until-dead discipline;
+// on the SIGTERM fallback path the process is already gone, so the first ping
+// fails and the wait returns immediately. Returns an error when the daemon is
+// still answering at the deadline — the caller should warn and proceed.
+func WaitForShutdownCompletion() error {
+	deadline := time.Now().Add(shutdownCompleteGrace)
+	for time.Now().Before(deadline) {
+		if pingDaemon() != nil {
+			return nil
+		}
+		time.Sleep(shutdownCompletePoll)
+	}
+	return fmt.Errorf("daemon control socket still answering %s after shutdown was acknowledged", shutdownCompleteGrace)
+}
+
+// isDaemonAbsentErr reports whether err from a dial/RPC call indicates that
+// no daemon is currently listening on the control socket (vs. some other
+// transport failure). Both ECONNREFUSED (stale socket, no listener) and
+// ENOENT (socket removed between Stat and Dial) qualify. Application-level
+// RPC errors (method-not-found, server panic) do NOT — those are handled
+// separately by isRPCMethodNotFoundErr so we can route them to the SIGTERM
+// fallback rather than treating them as "no daemon" and silently leaving the
+// stale process running (#504).
+func isDaemonAbsentErr(err error) bool {
+	if err == nil {
+		return false
+	}
+	if errors.Is(err, syscall.ECONNREFUSED) || errors.Is(err, fs.ErrNotExist) {
+		return true
+	}
+	return false
+}
+
+// isRPCMethodNotFoundErr reports whether err is the net/rpc server's reply
+// for an unknown method or service. The connection succeeded (daemon is
+// running, control socket is alive) but the registered service did not have
+// the requested method — i.e. a pre-#501 daemon that never registered
+// "Control.Shutdown". The stdlib returns this as rpc.ServerError with the
+// literal prefix "rpc: can't find method " or "rpc: can't find service ".
+func isRPCMethodNotFoundErr(err error) bool {
+	if err == nil {
+		return false
+	}
+	var serverErr rpc.ServerError
+	if !errors.As(err, &serverErr) {
+		return false
+	}
+	s := string(serverErr)
+	return strings.Contains(s, "can't find method") || strings.Contains(s, "can't find service")
+}
