+++ daemon/manager_create.go
@@ -1,0 +1,285 @@
+package daemon
+
+import (
+	"fmt"
+	"strings"
+
+	"github.com/sachiniyer/agent-factory/config"
+	"github.com/sachiniyer/agent-factory/session"
+	"github.com/sachiniyer/agent-factory/session/git"
+	"github.com/sachiniyer/agent-factory/session/tmux"
+	"github.com/sachiniyer/agent-factory/task"
+)
+
+func (m *Manager) CreateSession(req CreateSessionRequest) (session.InstanceData, error) {
+	if req.Program == "" {
+		// Default from the repo-resolved config so an in-repo
+		// default_program applies to daemon-created sessions (task runs,
+		// API creates) too. Falls back to the daemon's global config when
+		// the repo path can't be resolved — reserveCreate will surface
+		// that error with more context right after.
+		req.Program = m.cfg.DefaultProgram
+		if req.RepoPath != "" {
+			if repo, err := config.RepoFromPath(req.RepoPath); err == nil {
+				if resolved, rerr := config.ResolveConfig(repo.Root); rerr == nil {
+					req.Program = resolved.DefaultProgram
+				}
+			}
+		}
+	}
+	repo, title, release, err := m.reserveCreate(req)
+	if err != nil {
+		return session.InstanceData{}, err
+	}
+	defer release()
+
+	repoStartLock := m.startLockForRepo(repo.ID)
+	repoStartLock.Lock()
+	defer repoStartLock.Unlock()
+
+	instance, err := session.NewInstance(session.InstanceOptions{
+		Title:       title,
+		Path:        repo.Root,
+		Program:     req.Program,
+		AutoYes:     req.AutoYes,
+		InPlace:     req.InPlace,
+		ForceRemote: req.ForceRemote,
+	})
+	if err != nil {
+		return session.InstanceData{}, err
+	}
+
+	conversationCapture := session.BeginConversationCapture()
+
+	// Single creation flow (#930 PR 3): every instance owns its worktree 1:1.
+	// InPlace only changes WHICH worktree that is — the repo's own working tree,
+	// marked external — not the flow itself. finishCreateStart marks the instance
+	// live, PARKS it at a usage-limit wall (#1146 PR4), or returns a fatal error.
+	if serr := finishCreateStart(instance, req.Prompt, task.StartAndSendPrompt(instance, req.Prompt)); serr != nil {
+		_ = instance.Kill()
+		return session.InstanceData{}, fmt.Errorf("failed to start instance: %w", serr)
+	}
+	data := instance.ToInstanceData()
+
+	// Register the in-memory instance and persist it to disk inside the
+	// same critical section. The daemon refresh loop rebuilds
+	// session.Instance objects from disk for any key it does not already
+	// see in m.instances, so a window where the entry exists on disk but
+	// not in memory would let refresh construct a duplicate Instance
+	// (opening a fresh PTY in the tmux backend) that gets orphaned when
+	// the original is later stored under the same key.
+	key := daemonInstanceKey(repo.ID, title)
+	persistErr := func() error {
+		m.mu.Lock()
+		defer m.mu.Unlock()
+		m.instances[key] = instance
+		if err := appendInstanceData(repo.ID, data); err != nil {
+			delete(m.instances, key)
+			return err
+		}
+		return nil
+	}()
+	if persistErr != nil {
+		_ = instance.Kill()
+		return session.InstanceData{}, persistErr
+	}
+	m.captureAgentConversationAsync(repo.ID, key, instance, conversationCapture)
+
+	return data, nil
+}
+
+func (m *Manager) reserveCreate(req CreateSessionRequest) (*config.RepoContext, string, func(), error) {
+	if req.RepoPath == "" {
+		return nil, "", nil, fmt.Errorf("repo path is required")
+	}
+	repo, err := config.RepoFromPath(req.RepoPath)
+	if err != nil {
+		return nil, "", nil, err
+	}
+
+	m.mu.Lock()
+	defer m.mu.Unlock()
+	if err := m.refreshLocked(); err != nil {
+		return nil, "", nil, err
+	}
+
+	diskData, err := loadRepoInstanceData(repo.ID)
+	if err != nil {
+		return nil, "", nil, err
+	}
+
+	title := req.Title
+	if title == "" {
+		base := req.TitleBase
+		if base == "" {
+			return nil, "", nil, fmt.Errorf("session title is required")
+		}
+		title, err = m.nextAvailableTitleLocked(repo.ID, repo.Root, base, req.Program, req.ForceRemote, diskData)
+		if err != nil {
+			return nil, "", nil, err
+		}
+	} else if err := m.validateTitleAvailableLocked(repo.ID, repo.Root, title, req.Program, req.ForceRemote, req.allowReserved, diskData); err != nil {
+		return nil, "", nil, err
+	}
+
+	key := daemonInstanceKey(repo.ID, title)
+	remoteName := ""
+	if req.ForceRemote {
+		remoteName = session.Slugify(title)
+		if _, ok := m.reservedRemoteNames[remoteName]; ok {
+			return nil, "", nil, fmt.Errorf("remote hook name %q is already reserved", remoteName)
+		}
+	}
+
+	m.reservedTitles[key] = struct{}{}
+	if remoteName != "" {
+		m.reservedRemoteNames[remoteName] = struct{}{}
+	}
+	release := func() {
+		m.mu.Lock()
+		defer m.mu.Unlock()
+		delete(m.reservedTitles, key)
+		if remoteName != "" {
+			delete(m.reservedRemoteNames, remoteName)
+		}
+	}
+
+	return repo, title, release, nil
+}
+
+func (m *Manager) nextAvailableTitleLocked(repoID, repoPath, baseTitle, program string, remote bool, diskData []session.InstanceData) (string, error) {
+	for i := 1; i <= 10000; i++ {
+		candidate := baseTitle
+		if i > 1 {
+			candidate = fmt.Sprintf("%s-%d", baseTitle, i)
+		}
+		if err := m.validateTitleAvailableLocked(repoID, repoPath, candidate, program, remote, false, diskData); err == nil {
+			return candidate, nil
+		}
+	}
+	return "", fmt.Errorf("could not find an available title for %q", baseTitle)
+}
+
+func (m *Manager) validateTitleAvailableLocked(repoID, repoPath, title, program string, remote, allowReserved bool, diskData []session.InstanceData) error {
+	// Whitespace-only titles (e.g. "   ") are non-empty and so slip past a bare
+	// == "" check, creating sessions with effectively blank names (#973). Trim
+	// before the emptiness gate; the TUI naming flow applies the same check.
+	if strings.TrimSpace(title) == "" {
+		return fmt.Errorf("session title is required")
+	}
+	// The "root" title belongs to the daemon-managed root agent (#1106).
+	// Every creation path lands here — TUI, `af sessions create`, task
+	// spawns, DeliverPrompt auto-creates — so this single gate reserves the
+	// name everywhere. Only the daemon's own ensure loop passes
+	// allowReserved; title-base derivation (nextAvailableTitleLocked) never
+	// does, so a base of "root" skips to "root-2" instead of erroring.
+	if !allowReserved && session.IsReservedTitle(title) {
+		return fmt.Errorf("session title %q is reserved for the daemon-managed root agent; pick another name (to run a root agent on this repo, add it to root_agents in ~/.agent-factory/config.json)", title)
+	}
+	// Titles are sanitized into git branch names (git.SanitizeBranchName
+	// lowercases, turns spaces into dashes, strips unsafe chars, and collapses
+	// dashes), so distinct titles can map to the same branch: "MyApp"/"myapp"
+	// (#605) or "A B"/"a-b" (#741) both collide. The second worktree create
+	// would otherwise fail with a cryptic git error, so reject the conflict
+	// here, before any worktree or tmux setup runs.
+	if existing, kind := m.findTitleConflictLocked(repoID, title, diskData); existing != "" {
+		switch {
+		case existing == title:
+			if kind == titleConflictReserved {
+				return fmt.Errorf("session with title %q is already reserved: %w", title, errConcurrentCreate)
+			}
+			return fmt.Errorf("session with title %q already exists: %w", title, errConcurrentCreate)
+		default:
+			return fmt.Errorf("session titled %q conflicts with existing session %q: both sanitize to the same git branch %q", title, existing, m.branchForTitle(title))
+		}
+	}
+	if remote {
+		candidate := session.Slugify(title)
+		if _, ok := m.reservedRemoteNames[candidate]; ok {
+			return fmt.Errorf("remote hook name %q is already reserved", candidate)
+		}
+		for _, data := range diskData {
+			if data.BackendType != "remote" {
+				continue
+			}
+			if session.RemoteHookName(data.Title, data.RemoteMeta) == candidate {
+				return fmt.Errorf("remote session titled %q already maps to hook name %q", data.Title, candidate)
+			}
+		}
+		return nil
+	}
+	if tmuxSession := tmux.NewTmuxSessionForRepo(title, repoPath, program); tmuxSession.DoesSessionExist() {
+		// A tmux session exists with no daemon reservation, in-memory instance,
+		// or disk record — an orphan left by a crash or an external process.
+		// No creator will ever finish it, so this stays a plain error (not
+		// errConcurrentCreate): DeliverPrompt must fail fast with cleanup
+		// guidance rather than wait out waitForTargetSession's timeout (#916).
+		return fmt.Errorf("conflicting tmux session %q is already running; no agent-factory session owns it. Clean it up with: tmux kill-session -t %s", title, tmuxSession.SanitizedName())
+	}
+	return nil
+}
+
+type titleConflictKind int
+
+const (
+	titleConflictNone titleConflictKind = iota
+	titleConflictReserved
+	titleConflictLive
+	titleConflictDisk
+)
+
+// findTitleConflictLocked returns the existing title that conflicts with the
+// given candidate, along with the source of the conflict. An empty result means
+// the title is available. Two titles conflict when they derive the same git
+// branch name: branches are produced by git.SanitizeBranchName, which lowercases
+// and normalizes (spaces -> dashes, unsafe chars stripped, dashes collapsed),
+// so distinct titles like "MyApp"/"myapp" (#605) or "A B"/"a-b" (#741) can map
+// to one branch. Rejecting the collision here keeps the second worktree create
+// from failing with a cryptic git error.
+func (m *Manager) findTitleConflictLocked(repoID, title string, diskData []session.InstanceData) (string, titleConflictKind) {
+	for key := range m.reservedTitles {
+		rid, existing := splitDaemonInstanceKey(key)
+		if rid == repoID && m.titlesCollide(existing, title) {
+			return existing, titleConflictReserved
+		}
+	}
+	for key, inst := range m.instances {
+		rid, _ := splitDaemonInstanceKey(key)
+		if rid != repoID || inst == nil {
+			continue
+		}
+		if m.titlesCollide(inst.Title, title) {
+			return inst.Title, titleConflictLive
+		}
+	}
+	for _, data := range diskData {
+		if !m.titlesCollide(data.Title, title) {
+			continue
+		}
+		// Loading entries are transient TUI state with an empty worktree
+		// path and cannot be restored. Older TUI binaries (#551) could
+		// persist them to disk on quit, where they would block title
+		// reuse forever. Treat them as ghosts that the next save will
+		// reap rather than as live reservations.
+		if data.Status == session.Loading {
+			continue
+		}
+		return data.Title, titleConflictDisk
+	}
+	return "", titleConflictNone
+}
+
+// titlesCollide reports whether two session titles cannot coexist in the same
+// repo because they would derive the same git branch. It delegates to the shared
+// git.TitlesCollide helper so the daemon's authoritative validation and the
+// TUI's naming pre-check stay in lockstep (#936).
+func (m *Manager) titlesCollide(a, b string) bool {
+	return git.TitlesCollide(a, b, m.cfg.BranchPrefix)
+}
+
+// branchForTitle derives the git branch name for a session title using the same
+// prefix and sanitization the git worktree layer applies, so the daemon can
+// detect branch collisions before worktree setup runs.
+func (m *Manager) branchForTitle(title string) string {
+	return git.BranchForTitle(m.cfg.BranchPrefix, title)
+}
